Iterate with foreach and for
In pa.scala, we are passing in a function literal that takes one
parameter named arg. The body of the funciton is println(arg)

In epa.scala, we are passing in a function literal explicitly.

If an function literal consists of one statement that takes a single
argument, we need not explicitly name and specify the argument. For
example, we can do this:
args.foreach(println)