Parameterize arrays with types
What is Array[String](3)?
It is an array of string that is initialized to length 3 by passing the
value 3 to a constructor in parentheses.

The type of greetStrings is Array[String], not Array[String](3)

When we define a variable with val, the variable can't be reassigned,
but the object to which it refers could potentially still be mutated.