Understand classes
In class SimpleGreeter, greeting is a field and greet is a method. To
use the class, we initialize a val named g with a new instance of
SimpleGreeter. We then invoke the greet instance method on g.

In class FancyGreeter, greeting essentially becomes a private val (not a
var - it can't be reassigned) field that's accessible anywhere inside
the class body, but not accessible outside.

In class CarefulGreeter, we check the String used to parameterize a new
FancyGreeter instance for null, and throw NullPointerException to abort
the construction of the new instance.

In RepeatGreeter, "def this" is an additional auxiliary constructor that
takes a single String greeting parameter.