Understand singleton objects
Scala can't have any static fields or methods in a Scala class. Instead,
Scala allows we to create singleton objects using the keyword object. A
singleton object can not, and need not, be instantiated with new.

A singleton object is essentially automatically instantiated the first
time it is used.

A singleton object can share the same name with a class, and when it
does, the singleton is called the class's companion object.

* How to use the companion object's method?
Use the companion object name, a dot, and the method name

In WorldlyApp.scala, there is no class named WorldlyApp, so this singleton
object is not a companion object. It is instead called a stand-alone
object.

Neither WorldlyGreeter.scala nor WordlyApp.scala are scripts, because
they end in a definition.

For further understanding, read page 85 from programming in scala.