Method overloading and going further
To make Rational even more convenient, we can add new methods to the
class that perform mixed arithmetic on rationals and integers.

Now, there are two versions of each arithmetic operator method: one that
takes a rational as argument, the other which takes an integer.

We can create an implicit conversion which automatically converts
integers to rationals when needed.

=> implicit def intToRational(x: Int) = new Rational(x)

This defines a conversion method from Int to Rational. The implicit
modifier in front of the method tells the compiler to apply it
automatically in a number of situations.